Project/DjangoRESTfulMemo/requirements.txt
Django==2.0.13
django-cors-headers==3.1.1
djangorestframework==3.10.3
pkg-resources==0.0.0
pytz==2019.3
